Fix data corruption, body size limit, fragile IDs, and duplicated validation (#17-#20)
- Deep copy via JSON round-trip in storage copyMap, fix Patch to copy incoming data (closes #17)
- Add http.MaxBytesReader 1MB cap to Create/Update/Patch handlers (closes #18)
- Replace hand-rolled formatID/parseIDNumber with strconv stdlib calls (closes #19)
- Extract shared ValidateFieldType to pkg/types, remove duplicate in server and schema (closes #20)

// ticktockbent

+++ internal/storage/storage_test.go
@@ -260,7 +260,7 @@ func TestPatch(t *testing.T) {
 	store.Initialize([]string{"users"})
 
 	// Create a test entity
-	testData := map[string]interface{}{"name": "Alice", "email": "alice@example.com", "age": 30}
+	testData := map[string]interface{}{"name": "Alice", "email": "alice@example.com", "age": float64(30)}
 	id, _ := store.Create("users", testData)
 
 	tests := []struct {
@@ -274,7 +274,7 @@ func TestPatch(t *testing.T) {
 			name:       "patch existing entity",
 			entityType: "users",
 			id:         id,
-			data:       map[string]interface{}{"age": 31}, // Only update age
+			data:       map[string]interface{}{"age": float64(31)}, // Only update age
 			wantErr:    false,
 		},
 		{
